Fixing nested waits & cannot resolve > once
This commit updates promises so that they can only be resolved or
rejected once, and instead of ignoring subsequent attempts, throws an
exception. This is to make it easier to spot bugs related to promise
resolution. This commit updates the wait function to wait on a result
until a value is returned, including waiting on any promises (i.e., if
you're waiting on a then promise that returns a promise, then the wait
function will also wait on the forwarded promise).