Adding coroutine promises.
This commit adds coroutine promises similar to bluebird's
Promise.coroutine(). This allows you to create a promise that wraps a
coroutine that yields promises. The end result is the ability to write
code that interacts with promises in an iterative way.

Also added the exception_for function to create an exception for a
rejected promise reason.