Properly waiting on nested promise resolution
This commit updates the wait function to properly wait on nested
promises (when a promise is resolved with a promise). The actual
return/throw of a waited upon promise should be based on the last
resolved promise. For example, if a promise is rejected with a promise,
then the wait function should unwrap the state of the rejected promise
rather than the state of the promise that was rejected (allowing for
things like handling the rejection down the chain and ultimately
returning a fulfilled value).

Exceptions thrown when waiting on a promise after the promise is
fulfilled will now be rethrown as this points to an application error.
You will need to try/catch these types of wait functions in order to
work around this.