Fix manually settle promises generated with Utils::task
Promises generated with Utils::task throw a error when is canceled ou resolve before TaskQueue:run is called.
With this behavior, we can't cancel a task generated with Utils::task method, a very useful behavior.
In this fix the waitFn function generated in Utils:task is checking if promise is alredy settle, and skip the call to task function in this case.